Modified project structure to allow for testing of matrix multiplication functionality present in labeledImage.  The functionality that allowed for the determination of a transform in projective space mapping one quadrilateral to another was moved to projectivetransform.cpp / projectivetransform.h .  The source files were split into src/ and test/ directories to allow for future easier testing of modules in this project.  A new test for projectiveTransform, of the name of test_projectiveTransform, was created. projectiveTransform does not still function as planned, but progress in getting it to properly function will be much accelerated now that tests are in place and now that it has been separated from the rest of the code.

+++ src/imagelabelingscene.cpp
@@ -220,7 +220,7 @@ void ImageLabelingScene::save()
                         tempSavedPoints[3].toPoint()
                         )
                     );
-        currentLabeledImage->setBbox(pair, labelingState/totalStates);
+        currentLabeledImage->setBbox(pair, int(labelingState/totalStates));
 
         QGraphicsPathItem *path = dynamic_cast<QGraphicsPathItem*>(gTempItems["path"]);
         QGraphicsPolygonItem *poly = new QGraphicsPolygonItem(path->path().toFillPolygon());
